Move all tests to be of the public API
