+++ src/sh.rs
@@ -187,92 +187,3 @@ fn escape_size(char: &Char) -> usize {
         Extended(_) => 4,
     }
 }
-
-// ----------------------------------------------------------------------------
-
-#[cfg(test)]
-mod tests {
-    use std::os::unix::prelude::OsStringExt;
-    use std::{ffi::OsString, os::unix::ffi::OsStrExt};
-
-    use crate::find_bins;
-
-    use super::Sh;
-
-    #[test]
-    fn test_lowercase_ascii() {
-        assert_eq!(
-            Sh::quote("abcdefghijklmnopqrstuvwxyz"),
-            b"abcdefghijklmnopqrstuvwxyz"
-        );
-    }
-
-    #[test]
-    fn test_uppercase_ascii() {
-        assert_eq!(
-            Sh::quote("ABCDEFGHIJKLMNOPQRSTUVWXYZ"),
-            b"ABCDEFGHIJKLMNOPQRSTUVWXYZ"
-        );
-    }
-
-    #[test]
-    fn test_numbers() {
-        assert_eq!(Sh::quote("0123456789"), b"0123456789");
-    }
-
-    #[test]
-    fn test_punctuation() {
-        assert_eq!(Sh::quote("-_=/,.+"), b"'-_=/,.+'");
-    }
-
-    #[test]
-    fn test_empty_string() {
-        assert_eq!(Sh::quote(""), b"''");
-    }
-
-    #[test]
-    fn test_basic_escapes() {
-        assert_eq!(Sh::quote(r#"woo'wah""#), br#"'woo\047wah"'"#);
-    }
-
-    #[test]
-    fn test_control_characters() {
-        assert_eq!(Sh::quote("\x07"), b"'\\a'");
-        assert_eq!(Sh::quote("\x00"), b"'\\000'");
-        assert_eq!(Sh::quote("\x06"), b"'\\006'");
-        assert_eq!(Sh::quote("\x7F"), b"'\x7F'");
-        assert_eq!(Sh::quote("\x1B"), b"'\\033'");
-    }
-
-    #[test]
-    fn test_quote_into_plain() {
-        let mut buffer = Vec::new();
-        Sh::quote_into("hello", &mut buffer);
-        assert_eq!(buffer, b"hello");
-    }
-
-    #[test]
-    fn test_quote_into_with_escapes() {
-        let mut buffer = Vec::new();
-        Sh::quote_into("-_=/,.+", &mut buffer);
-        assert_eq!(buffer, b"'-_=/,.+'");
-    }
-
-    #[test]
-    fn test_roundtrip() {
-        use std::process::Command;
-        // In Bash it doesn't seem possible to roundtrip NUL, but in the Bourne
-        // shell, or whatever is masquerading as `sh`, it seems to be fine.
-        let string: OsString = OsString::from_vec((u8::MIN..=u8::MAX).collect());
-        let mut script = b"echo ".to_vec();
-        Sh::quote_into(string.as_bytes(), &mut script);
-        let script = OsString::from_vec(script);
-        for bin in find_bins("sh") {
-            let output = Command::new(bin).arg("-c").arg(&script).output().unwrap();
-            let mut result = output.stdout;
-            result.resize(result.len() - 1, 0); // Remove trailing newline.
-            let result = OsString::from_vec(result);
-            assert_eq!(result, string);
-        }
-    }
-}
